+++ src/views/page-header.usage.js
@@ -1,11 +1,14 @@
 export default (view) => ({
-    demo: `${view}:"Page header \\"${view}\\""`,
+    demo: `${view}:"That's a \\"${view}\\""`,
     examples: [
         {
             title: 'Prelude',
             demo: {
                 view,
-                prelude: 'badge:{ text: "demo" }',
+                prelude: [
+                    'badge:{ text: "demo" }',
+                    'badge:{ text: "demo", prefix: "prelude", postfix: "postfix" }'
+                ],
                 content: 'h1:"Header"'
             }
         }